requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
numpy==1.24.3
sqlite-utils==3.35.2
python-multipart==0.0.6
aiofiles==23.2.1
transformers==4.36.0
torch==2.1.0